Edit page fixes
